feat: Add transaction form and modal
Introduces a new modal for adding transactions and integrates it into the DatabaseView. This allows users to manually add new income or expense entries.

The `AppContext` has been updated to include a `handleAddTransaction` function to manage the addition of new transactions to the application state.

A new `XIcon` component has also been added to the icons library for closing the modal.